Added a lightweight fp library
Most fp libraries include a whole slew of functions to operate on
arrays. However, JS already has most of these defined, making redefining
them pointless.

This library only includes functions that are not already defined in JS.
It is also intentionally incomplete and (given the size of most fp type
libraries) misses out a lot of functions that may be useful. The point
however, is that these functions are all really easy to define given the
set of functions defined here.